requirements.txt
transformers>=4.30.0
torch>=2.0.0
peft>=0.4.0
datasets>=2.12.0
streamlit>=1.24.0
huggingface-hub>=0.16.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
python-docx>=0.8.11
redis>=4.5.0
gunicorn>=20.1.0
nginx>=1.24.0
prometheus-client>=0.16.0
sentry-sdk>=1.25.0